move read/write_attribs to attrib.h, too
