+++ src/util/attrib.c
@@ -29,6 +29,36 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <string.h>
 #include <stdlib.h>
 
+int read_attribs(gamedata *data, attrib **alist, void *owner) {
+    int result;
+    if (data->version < ATHASH_VERSION) {
+        result = a_read_orig(data, alist, owner);
+    }
+    else {
+        result = a_read(data, alist, owner);
+    }
+    if (result == AT_READ_DEPR) {
+        /* handle deprecated attributes */
+        attrib *a = *alist;
+        while (a) {
+            if (a->type->upgrade) {
+                a->type->upgrade(alist, a);
+            }
+            a = a->nexttype;
+        }
+    }
+    return result;
+}
+
+void write_attribs(storage *store, attrib *alist, const void *owner)
+{
+#if RELEASE_VERSION < ATHASH_VERSION
+    a_write_orig(store, alist, owner);
+#else
+    a_write(store, alist, owner);
+#endif
+}
+
 int a_readint(attrib * a, void *owner, struct gamedata *data)
 {
     int n;
